+++ shacl_ir/src/compiled/severity.rs
@@ -2,6 +2,7 @@ use std::fmt::Display;
 
 use iri_s::IriS;
 use shacl_ast::shacl_vocab::{sh_info, sh_violation, sh_warning};
+use shacl_ast::{sh_debug, sh_trace};
 
 use shacl_ast::severity::Severity;
 
@@ -9,15 +10,19 @@ use super::compiled_shacl_error::CompiledShaclError;
 
 #[derive(Hash, Clone, PartialEq, Eq, Debug)]
 pub enum CompiledSeverity {
-    Violation,
-    Warning,
+    Trace,
+    Debug,
     Info,
+    Warning,
+    Violation,
     Generic(IriS),
 }
 
 impl CompiledSeverity {
     pub fn iri(&self) -> IriS {
         match self {
+            CompiledSeverity::Trace => sh_trace().clone(),
+            CompiledSeverity::Debug => sh_debug().clone(),
             CompiledSeverity::Violation => sh_violation().clone(),
             CompiledSeverity::Warning => sh_warning().clone(),
             CompiledSeverity::Info => sh_info().clone(),
@@ -29,6 +34,8 @@ impl CompiledSeverity {
         let ans = match severity {
             Some(severity) => {
                 let severity = match severity {
+                    Severity::Trace => CompiledSeverity::Trace,
+                    Severity::Debug => CompiledSeverity::Debug,
                     Severity::Violation => CompiledSeverity::Violation,
                     Severity::Warning => CompiledSeverity::Warning,
                     Severity::Info => CompiledSeverity::Info,
@@ -46,11 +53,27 @@ impl CompiledSeverity {
 
         Ok(ans)
     }
-}
 
-impl From<&CompiledSeverity> for IriS {
-    fn from(value: &CompiledSeverity) -> Self {
-        match value {
+    pub fn from_iri(iri: &IriS) -> Option<Self> {
+        if iri == sh_trace() {
+            Some(CompiledSeverity::Trace)
+        } else if iri == sh_debug() {
+            Some(CompiledSeverity::Debug)
+        } else if iri == sh_violation() {
+            Some(CompiledSeverity::Violation)
+        } else if iri == sh_warning() {
+            Some(CompiledSeverity::Warning)
+        } else if iri == sh_info() {
+            Some(CompiledSeverity::Info)
+        } else {
+            Some(CompiledSeverity::Generic(iri.clone()))
+        }
+    }
+
+    pub fn to_iri(&self) -> IriS {
+        match self {
+            CompiledSeverity::Trace => sh_trace().clone(),
+            CompiledSeverity::Debug => sh_debug().clone(),
             CompiledSeverity::Violation => sh_violation().clone(),
             CompiledSeverity::Warning => sh_warning().clone(),
             CompiledSeverity::Info => sh_info().clone(),
@@ -59,9 +82,17 @@ impl From<&CompiledSeverity> for IriS {
     }
 }
 
+impl From<&CompiledSeverity> for IriS {
+    fn from(value: &CompiledSeverity) -> Self {
+        CompiledSeverity::to_iri(value)
+    }
+}
+
 impl Display for CompiledSeverity {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
+            CompiledSeverity::Trace => write!(f, "Trace"),
+            CompiledSeverity::Debug => write!(f, "Debug"),
             CompiledSeverity::Violation => write!(f, "Violation"),
             CompiledSeverity::Warning => write!(f, "Warning"),
             CompiledSeverity::Info => write!(f, "Info"),